refactor(gsd): narrow worktree sync test seams
